+++ test/WebJobs.Script.Tests/FunctionMetadataManagerTests.cs
@@ -188,7 +188,43 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
         }
 
         [Fact]
-        public void FunctionMetadataManager_HandlesTimeoutsGracefully_WhenFunctionProvidersDoNotRespond()
+        public void FunctionMetadataManager_LoadFunctionMetadata_Throws_WhenFunctionProviderThrows()
+        {
+            var functionMetadataCollection = new Collection<FunctionMetadata>();
+            var mockFunctionErrors = new Dictionary<string, ImmutableArray<string>>();
+            var mockFunctionMetadataProvider = new Mock<IFunctionMetadataProvider>();
+            var badFunctionMetadataProvider = new Mock<IFunctionProvider>();
+            var goodFunctionMetadataProvider = new Mock<IFunctionProvider>();
+            var workerConfigs = TestHelpers.GetTestWorkerConfigs();
+            var testLoggerProvider = new TestLoggerProvider();
+            var loggerFactory = new LoggerFactory();
+            loggerFactory.AddProvider(testLoggerProvider);
+
+            mockFunctionMetadataProvider.Setup(m => m.GetFunctionMetadataAsync(workerConfigs, SystemEnvironment.Instance, false)).Returns(Task.FromResult(new Collection<FunctionMetadata>().ToImmutableArray()));
+            mockFunctionMetadataProvider.Setup(m => m.FunctionErrors).Returns(new Dictionary<string, ICollection<string>>().ToImmutableDictionary(kvp => kvp.Key, kvp => kvp.Value.ToImmutableArray()));
+
+            // A good provider that returns 2 functions
+            functionMetadataCollection.Add(GetTestFunctionMetadata("somefile.dll", name: "Function1"));
+            functionMetadataCollection.Add(GetTestFunctionMetadata("somefile.dll", name: "Function2"));
+            goodFunctionMetadataProvider.Setup(m => m.GetFunctionMetadataAsync()).ReturnsAsync(functionMetadataCollection.ToImmutableArray());
+
+            // A bad provider that will throw an exception for .GetFunctionMetadataAsync call.
+            var tcs = new TaskCompletionSource<ImmutableArray<FunctionMetadata>>();
+            badFunctionMetadataProvider.Setup(m => m.GetFunctionMetadataAsync()).Throws(new IOException("There was a custom IO error"));
+
+            FunctionMetadataManager testFunctionMetadataManager = TestFunctionMetadataManager.GetFunctionMetadataManager(new OptionsWrapper<ScriptJobHostOptions>(_scriptJobHostOptions),
+                mockFunctionMetadataProvider.Object, new List<IFunctionProvider>() { goodFunctionMetadataProvider.Object, badFunctionMetadataProvider.Object }, new OptionsWrapper<HttpWorkerOptions>(_defaultHttpWorkerOptions), loggerFactory, new TestOptionsMonitor<LanguageWorkerOptions>(TestHelpers.GetTestLanguageWorkerOptions()));
+
+            var exception = Assert.Throws<IOException>(() => testFunctionMetadataManager.LoadFunctionMetadata());
+            Assert.Contains($"There was a custom IO error", exception.Message);
+
+            var traces = testLoggerProvider.GetAllLogMessages();
+            Assert.Single(traces, t => t.FormattedMessage.Contains("Reading functions metadata (Custom)"));
+            Assert.DoesNotContain(traces, t => t.FormattedMessage.Contains("2 functions found (Custom)"));
+        }
+
+        [Fact]
+        public void FunctionMetadataManager_LoadFunctionMetadata_Throws_WhenFunctionProvidersTimesOut()
         {
             var functionMetadataCollection = new Collection<FunctionMetadata>();
             var mockFunctionErrors = new Dictionary<string, ImmutableArray<string>>();
@@ -210,7 +246,11 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
 
             // A bad provider that will never return anything.
             var tcs = new TaskCompletionSource<ImmutableArray<FunctionMetadata>>();
-            badFunctionMetadataProvider.Setup(m => m.GetFunctionMetadataAsync()).Returns(tcs.Task);
+            badFunctionMetadataProvider.Setup(m => m.GetFunctionMetadataAsync()).Returns(async () =>
+            {
+                await Task.Delay(TimeSpan.FromSeconds(2)); // Simulate a delay longer than the timeout
+                return ImmutableArray<FunctionMetadata>.Empty;
+            });
 
             FunctionMetadataManager testFunctionMetadataManager = TestFunctionMetadataManager.GetFunctionMetadataManager(new OptionsWrapper<ScriptJobHostOptions>(_scriptJobHostOptions),
                 mockFunctionMetadataProvider.Object, new List<IFunctionProvider>() { goodFunctionMetadataProvider.Object, badFunctionMetadataProvider.Object }, new OptionsWrapper<HttpWorkerOptions>(_defaultHttpWorkerOptions), loggerFactory, new TestOptionsMonitor<LanguageWorkerOptions>(TestHelpers.GetTestLanguageWorkerOptions()));
@@ -218,17 +258,12 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             // Set the timeout to 1 second for the test.
             testFunctionMetadataManager.MetadataProviderTimeoutInSeconds = 1;
 
-            // Run LoadFunctionMetadata in a separate Task to avoid blocking the test thread
-            var loadFunctionMetadataTask = Task.Run(() => testFunctionMetadataManager.LoadFunctionMetadata());
+            var exception = Assert.Throws<TimeoutException>(() => testFunctionMetadataManager.LoadFunctionMetadata());
+            Assert.Contains($"Timeout occurred while retrieving metadata from provider '{badFunctionMetadataProvider.Object.GetType().FullName}'. The operation exceeded the configured timeout of 1 seconds.", exception.Message);
 
-            // wait couple of seconds.
-            loadFunctionMetadataTask.Wait(TimeSpan.FromSeconds(3));
             var traces = testLoggerProvider.GetAllLogMessages();
-            Assert.Equal(1, traces.Count(t => t.FormattedMessage.Contains("Reading functions metadata (Custom)")));
-            // We should see log entry for the good provider.
-            Assert.Contains(traces, t => t.FormattedMessage.Contains("2 functions found (Custom)"));
-            // We should see an error log entry for the bad provider.
-            Assert.Contains(traces, t => t.FormattedMessage.Contains($"Timeout or failure in retrieving metadata from '{badFunctionMetadataProvider.Object.GetType().FullName}'"));
+            Assert.Single(traces, t => t.FormattedMessage.Contains("Reading functions metadata (Custom)"));
+            Assert.DoesNotContain(traces, t => t.FormattedMessage.Contains("2 functions found (Custom)"));
         }
 
         [Fact]